CA-17200: Introduce answerfile support for OEM-HDD edition
Answerfile should be invoked with these on the kernel command line:
--answerfile
--install_oem_to_disk
--reboot

Invoking the installer at the command line:

python /opt/xensource/installer/init --oem_edition --OEM_MANUFACTURER=Citrix --install_oem_to_disk --answerfile=nfs://192.168.1.99:/home/foo/export/answers --reboot

Sample answerfile that will read the OEM image from local media on sr0:

<install mode="oemhdd" srtype="lvm">
<primary-disk>sda</primary-disk>
<source type="local">sr0:oem-citrix-graphical.img.bz2</source>
</install>

Sample answerfile that will read the OEM image from local media on sda2:

<install mode="oemhdd" srtype="lvm">
<primary-disk>sda</primary-disk>
<source type="local">sda2:/path/to/image/oem-citrix-graphical.img.bz2</source>
</install>

Sample answerfile that will scan local partitions and media devices for OEM image
with the specified filename (full path filename expected unless file in / dir):

<install mode="oemhdd" srtype="lvm">
<primary-disk>sda</primary-disk>
<source type="local">oem-citrix-graphical.img.bz2</source>
</install>

Sample answerfile that will retrieve OEM image via nfs:

<install mode="oemhdd" srtype="lvm">
<primary-disk>sda</primary-disk>
<source type="nfs">192.168.1.99:/home/foo/export/oem-citrix-graphical.img.bz2</source>
</install>

Sample answerfile that will retrieve OEM image via nfs, and boot with XenRT
specified NIC enabled:

<install mode="oemhdd" srtype="lvm">
<primary-disk>sda</primary-disk>
<source type="local">oem-citrix-graphical.img.bz2</source>
<xenrt>eth0</xenrt>
</install>

Sample answerfile that will retrieve OEM image via nfs, and boot with XenRT enabled
and disks scorched as per reset-to-factory

<install mode="oemhdd" srtype="lvm">
<primary-disk>sda</primary-disk>
<source type="local">oem-citrix-graphical.img.bz2</source>
<xenrt scorch="true">eth0</xenrt>
</install>

+++ oem.py
@@ -28,6 +28,7 @@ import product
 import diskutil
 import init_constants
 import xelogging
+import answerfile
 import netutil
 import util
 import repository
@@ -59,11 +60,12 @@ def writeImageWithProgress(ui, devnode, answers):
     reads_done = 0
     reads_needed = int(float(bzfilesize)/float(rdbufsize))
 
-    pd = ui.progress.initProgressDialog(
-        "Decompressing image",
-        "%(image_name)s is being written to %(devnode)s" % locals(),
-        reads_needed)
-    ui.progress.displayProgressDialog(0, pd)
+    if ui:
+        pd = ui.progress.initProgressDialog(
+            "Decompressing image",
+            "%(image_name)s is being written to %(devnode)s" % locals(),
+            reads_needed)
+        ui.progress.displayProgressDialog(0, pd)
 
     try:
         while True:
@@ -72,29 +74,138 @@ def writeImageWithProgress(ui, devnode, answers):
             if not buffer:
                 break
             devfd.write(decompressor(buffer))
-            ui.progress.displayProgressDialog(min(reads_needed, reads_done), pd)
+            if ui:
+                ui.progress.displayProgressDialog(min(reads_needed, reads_done), pd)
     except Exception, e:
         xelogging.log_exception(e)
-        ui.progress.clearModelessDialog()
+        if ui:
+            ui.progress.clearModelessDialog()
         image_fd.close()
         devfd.close()
-        ui.OKDialog("Error", "Fatal error occurred during write.  Press any key to reboot")
+        if ui:
+            ui.OKDialog("Error", "Fatal error occurred during write.  Press any key to reboot")
         return EXIT_ERROR
 
     # image successfully written
-    ui.progress.clearModelessDialog()
+    if ui:
+        ui.progress.clearModelessDialog()
     image_fd.close()
     devfd.close()
 
     return EXIT_OK
 
-# TODO answerfile support - see install.go
-def go_disk(ui, args):
+def post_process_answerfile_data(results):
+    "Processing the answerfile entries to derive data"
+
+    dirname  = os.path.dirname(results['source-address'])
+    basename = os.path.basename(results['source-address'])
+
+    results['image-name'] = basename
+
+    if results['source-media'] == 'local':
+
+        if len(dirname.split(':')) > 1:
+            (dev, dirname) = dirname.split(':', 1)
+            device_path = "/dev/%s" % dev
+            if not os.path.exists(device_path):
+                # Device path doesn't exist (maybe udev renamed it).  Create it now.
+                major, minor = map(int, open('/sys/block/%s/dev' % dev).read().split(':'))
+                os.mknod(device_path, 0600|stat.S_IFBLK, os.makedev(major,minor))
+
+            da = repository.DeviceAccessor(device_path)
+            try:
+                da.start()
+            except util.MountFailureException:
+                raise AnswerfileError, "Could not mount local %s to read image." % device_path
+
+            fulldirpath = da.location + '/' + dirname
+            fullpath = os.path.join(fulldirpath, basename)
+
+            # check for existence using stat
+            try:
+                imageSize = os.stat(fullpath).st_size
+            except OSError:
+                raise AnswerfileError, "No local OEM image found at %s." % results['source-address']
+
+            results['image-size'] = imageSize
+            results['image-fd'] = open(fullpath, "rb")
+            results['accessor'] = da
+
+        else:
+            # scan all local devices for the file
+            # *including* local disk partitions (think: factory install)
+            devs_and_ptns = diskutil.getPartitionList()
+            devs_and_ptns.extend(diskutil.getRemovableDeviceList())
+
+            for check in devs_and_ptns:
+                device_path = "/dev/%s" % check
+                if not os.path.exists(device_path):
+                    # Device path doesn't exist (maybe udev renamed it).  Create it now.
+                    major, minor = map(int, open('/sys/block/%s/dev' % check).read().split(':'))
+                    os.mknod(device_path, 0600|stat.S_IFBLK, os.makedev(major,minor))
+
+                da = repository.DeviceAccessor(device_path)
+                try:
+                    da.start()
+                except util.MountFailureException:
+                    continue
+
+                fulldirpath = os.path.join(da.location, dirname)
+                fullpath = os.path.join(fulldirpath, basename)
+                try:
+                    imageSize = os.stat(fullpath).st_size
+                except OSError:
+                    da.finish()
+                    continue
+
+                results['image-size'] = imageSize
+                results['image-fd'] = open(fullpath, "rb")
+                results['accessor'] = da
+                break
+
+            if not results.has_key('accessor'):
+                raise AnswerfileError, "Scan found no local OEM image at %s." % results['source-address']
+    else:
+        if results['source-media'] == 'nfs':
+            accessor = repository.NFSAccessor(dirname)
+        else:
+            accessor = repository.URLAccessor(dirname)
+
+        try:
+            accessor.start()
+        except:
+            raise AnswerfileError, "Could not reach image at %s." % results['source-address']
+
+        if not accessor.access(basename):
+            accessor.finish()
+            raise AnswerfileError, "Could not find image at %s." % basename
+
+        results['image-fd'] = accessor.openAddress(basename)
+        results['accessor'] = accessor
+
+        if results['source-media'] == 'nfs':
+            fullpath = os.path.join(accessor.location, basename)
+            results['image-size'] = os.stat(fullpath).st_size
+        else:
+            results['image-size'] = 900000000 # FIXME: A GUESS!
+
+
+def go_disk(ui, args, answerfile_address):
     "Install oem edition to disk"
-    xelogging.log("Starting install to disk dialog")
-    answers = ui.init_oem.recover_disk_drive_sequence()
-    if not answers:
-        return None # keeps outer loop going
+
+    # loading an answerfile?
+    assert ui != None or answerfile_address != None
+
+    if answerfile_address:
+        answers = answerfile.processAnswerfile(answerfile_address)
+        post_process_answerfile_data(answers)
+
+    else:
+        xelogging.log("Starting install to disk dialog")
+        if ui:
+            answers = ui.init_oem.recover_disk_drive_sequence()
+        if not answers:
+            return None # keeps outer loop going
 
     xelogging.log("Starting install to disk, partitioning")
 
@@ -103,17 +214,20 @@ def go_disk(ui, args):
     STATE_PARTITION_NUMBER = 7
     SR_PARTITION_NUMBER    = 8
 
-    devnode = "/dev/" + answers["blockdev"]
+    devnode = answers["primary-disk"]
 
     # Step 1: create system partitions.
     #
-    ui.progress.showMessageDialog("Partitioning", "Creating the system image disk partitions ...")
+    if ui:
+        ui.progress.showMessageDialog("Partitioning", "Creating the system image disk partitions ...")
     
     rv, output = util.runCmd('%s/create-partitions %s 2>&1' % (scriptdir,devnode), with_output=True)
-    ui.progress.clearModelessDialog()
+    if ui:
+        ui.progress.clearModelessDialog()
     if rv:
-        ui.OKDialog ("Error", "Fatal error occurred during disk partitioning:\n\n%s\n\n"
-                     "Press any key to reboot" % output)
+        if ui:
+            ui.OKDialog ("Error", "Fatal error occurred during disk partitioning:\n\n%s\n\n"
+                         "Press any key to reboot" % output)
         return EXIT_ERROR
 
     # Step 2: ensure that the device nodes for the new partitions are available
@@ -122,8 +236,9 @@ def go_disk(ui, args):
     # lookup the boot partition number
     rv, output = util.runCmd('/sbin/fdisk -l %s 2>&1 | /bin/sed -ne \'s,^%sp\\?\\([1-4]\\)  *\\*.*$,\\1,p\' 2>&1' % (devnode, devnode), with_output=True)
     if rv:
-        ui.OKDialog ("Error", "Fatal error identifying boot partition:\n\n%s\n\n"
-                     "Press any key to reboot" % output)
+        if ui:
+            ui.OKDialog ("Error", "Fatal error identifying boot partition:\n\n%s\n\n"
+                         "Press any key to reboot" % output)
         return EXIT_ERROR
     BOOT_PARTITION_NUMBER = int(output)
 
@@ -166,102 +281,124 @@ def go_disk(ui, args):
     xelogging.log("Populating system partitions")
 
     ###########################################################################
-    ui.progress.showMessageDialog("Imaging", "Installing master boot record...")
+    if ui:
+        ui.progress.showMessageDialog("Imaging", "Installing master boot record...")
     
     rv, output = util.runCmd('%s/populate-partition %s %s master-boot-record 2>&1' % (scriptdir,sr_devnode, devnode), with_output=True)
-    ui.progress.clearModelessDialog()
+    if ui:
+        ui.progress.clearModelessDialog()
     if rv:
-        ui.OKDialog ("Error", "Fatal error occurred during installation of master boot record:\n\n%s\n\n" 
-                     "Press any key to reboot" % output)
+        if ui:
+            ui.OKDialog ("Error", "Fatal error occurred during installation of master boot record:\n\n%s\n\n" 
+                         "Press any key to reboot" % output)
         return EXIT_ERROR
 
     ###########################################################################
-    ui.progress.showMessageDialog("Imaging", "Populating primary system image...")
+    if ui:
+        ui.progress.showMessageDialog("Imaging", "Populating primary system image...")
     
     partnode = getPartitionNode(devnode, SYS_1_PARTITION_NUMBER)
     rv, output = util.runCmd('%s/populate-partition %s %s system-image-1 2>&1' % (scriptdir,sr_devnode, partnode), with_output=True)
-    ui.progress.clearModelessDialog()
+    if ui:
+        ui.progress.clearModelessDialog()
     if rv:
-        ui.OKDialog ("Error", "Fatal error occurred during population of primary system image:\n\n%s\n\n" 
-                     "Press any key to reboot" % output)
+        if ui:
+            ui.OKDialog ("Error", "Fatal error occurred during population of primary system image:\n\n%s\n\n" 
+                         "Press any key to reboot" % output)
         return EXIT_ERROR
 
     ###########################################################################
-    ui.progress.showMessageDialog("Imaging", "Populating secondary system image...")
+    if ui:
+        ui.progress.showMessageDialog("Imaging", "Populating secondary system image...")
     
     partnode = getPartitionNode(devnode, SYS_2_PARTITION_NUMBER)
     rv, output = util.runCmd('%s/populate-partition %s %s system-image-2 2>&1' % (scriptdir,sr_devnode, partnode), with_output=True)
-    ui.progress.clearModelessDialog()
+    if ui:
+        ui.progress.clearModelessDialog()
     if rv:
-        ui.OKDialog ("Error", "Fatal error occurred during population of secondary system image:\n\n%s\n\n" 
-                     "Press any key to reboot" % output)
+        if ui:
+            ui.OKDialog ("Error", "Fatal error occurred during population of secondary system image:\n\n%s\n\n" 
+                         "Press any key to reboot" % output)
         return EXIT_ERROR
 
     ###########################################################################
-    ui.progress.showMessageDialog("Imaging", "Initialising writable storage...")
+    if ui:
+        ui.progress.showMessageDialog("Imaging", "Initialising writable storage...")
     
     partnode = getPartitionNode(devnode, STATE_PARTITION_NUMBER)
     rv, output = util.runCmd('%s/populate-partition %s %s mutable-state 2>&1' % (scriptdir,sr_devnode, partnode), with_output=True)
-    ui.progress.clearModelessDialog()
+    if ui:
+        ui.progress.clearModelessDialog()
     if rv:
-        ui.OKDialog ("Error", "Fatal error occurred initialising writable storage:\n\n%s\n\n" 
-                     "Press any key to reboot" % output)
+        if ui:
+            ui.OKDialog ("Error", "Fatal error occurred initialising writable storage:\n\n%s\n\n" 
+                         "Press any key to reboot" % output)
         return EXIT_ERROR
 
     ###########################################################################
-    ui.progress.showMessageDialog("Imaging", "Initialising boot partition...")
+    if ui:
+        ui.progress.showMessageDialog("Imaging", "Initialising boot partition...")
     
     partnode = getPartitionNode(devnode, BOOT_PARTITION_NUMBER)
     rv, output = util.runCmd('%s/populate-partition %s %s boot 2>&1' % (scriptdir,sr_devnode, partnode), with_output=True)
-    ui.progress.clearModelessDialog()
+    if ui:
+        ui.progress.clearModelessDialog()
     if rv:
-        ui.OKDialog ("Error", "Fatal error occurred initialising boot partition:\n\n%s\n\n" 
-                     "Press any key to reboot" % output)
+        if ui:
+            ui.OKDialog ("Error", "Fatal error occurred initialising boot partition:\n\n%s\n\n" 
+                         "Press any key to reboot" % output)
         return EXIT_ERROR
 
     ###########################################################################
-    ui.progress.showMessageDialog("VM Storage", "Creating disk partition for local storage...")
+    if ui:
+        ui.progress.showMessageDialog("VM Storage", "Creating disk partition for local storage...")
     
     # Create a partition containing the remaining disk space and tell XAPI to 
     # initialise it as the Local SR on first boot
     partnode = getPartitionNode(devnode, STATE_PARTITION_NUMBER)
     rv, output = util.runCmd('%s/update-partitions %s %s %s 2>&1' % (scriptdir,devnode,sr_devnode,partnode), with_output=True)
-    ui.progress.clearModelessDialog()
+    if ui:
+        ui.progress.clearModelessDialog()
     if rv:
-        ui.OKDialog ("Error", "Fatal error occurred during SR initialisation:\n\n%s\n\n" 
-                     "Press any key to reboot" % output)
+        if ui:
+            ui.OKDialog ("Error", "Fatal error occurred during SR initialisation:\n\n%s\n\n" 
+                         "Press any key to reboot" % output)
         return EXIT_ERROR
 
     ###########################################################################
     # update the initrds on the bootable partitions to support access to this disk
-    ui.progress.showMessageDialog("update-initrd", "Customising startup modules...")
+    if ui:
+        ui.progress.showMessageDialog("update-initrd", "Customising startup modules...")
     for part in (SYS_1_PARTITION_NUMBER, SYS_2_PARTITION_NUMBER):
         partnode = getPartitionNode(devnode,part)
         rv, output = util.runCmd('%s/update-initrd %s 2>&1' % (scriptdir,partnode), with_output=True)
         if rv:
             break;
     if rv:
-        ui.progress.clearModelessDialog()
-        ui.OKDialog ("Error", "Fatal error occurred during customisation of startup modules:\n\n%s\n\n" 
-                     "Press any key to reboot" % output)
+        if ui:
+            ui.progress.clearModelessDialog()
+            ui.OKDialog ("Error", "Fatal error occurred during customisation of startup modules:\n\n%s\n\n" 
+                         "Press any key to reboot" % output)
         return EXIT_ERROR
     
-    ui.progress.clearModelessDialog()
-
     # success!
-    ui.OKDialog("Success", "Install complete.  Click OK to reboot")
+    if ui:
+        ui.progress.clearModelessDialog()
+        ui.OKDialog("Success", "Install complete.  Click OK to reboot")
+
     return EXIT_OK
 
 # TODO answerfile support - see install.go
-def go_flash(ui, args):
+def go_flash(ui, args, answerfile_address):
     "Install oem edition to flash"
+
     xelogging.log("Starting install to flash dialog")
     answers = ui.init_oem.recover_pen_drive_sequence()
     if not answers:
         return None # keeps outer loop going
 
     xelogging.log("Starting install to flash write")
-    devnode = "/dev/" + answers["blockdev"]
+    devnode = answers["primary-disk"]
 
     rv = writeImageWithProgress(ui, devnode, answers)
     if rv == EXIT_OK:
